add Exceptions to the Service + a custom exception UserNotFoundException

+++ demo/src/main/java/com/project/demo/service/UserService.java
@@ -5,7 +5,11 @@ import com.project.demo.repository.UserRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.beans.factory.annotation.Autowired;
 import java.util.List;
+import org.springframework.dao.DataAccessException;
+import com.project.demo.exception.UserNotFoundException;
 
+
+/**possible DB exception should be auto-thrown from spring */
 
 @Service
 public class UserService {
@@ -18,21 +22,28 @@ public class UserService {
     }
 
     public List<User> findAll() {
-       return userRepository.findAll();
+            return userRepository.findAll();
     }
 
     public User findUser(Long id) {
-      return userRepository.findById(id).orElseThrow();
+
+            return userRepository.findById(id).orElseThrow(() -> new UserNotFoundException(id));
   
     }
 
     public User saveUser(User user) {
-       return userRepository.save(user);
+            return userRepository.save(user);
   
     }
 
-     public void deleteUser(Long id) {
-       userRepository.deleteById(id);
+    public void deleteUser(Long id) {
+
+        if (!userRepository.existsById(id)) {
+            throw new UserNotFoundException(id);
+        }
+
+            userRepository.deleteById(id);
+ 
   
     }
